# Enhance load_database Performance & Reduce Database File Size