## Prerequisite
- This PR depends on https://github.com/ProjectTorreyPines/FusionMaterials.jl/pull/13

## Summary
- **Use cached `ALL_MATERIALS` instead of `all_materials()`**
  - This change significantly improves the performance of loading `ini`
  
- **New keyword `sub_format` for `sample_and_write_database` function**
  - If `sub_format=:json`, it samples and writes the new database with `dd.json`, `ini.json`, and `act.json`.
  - Using `JSON` for subgroups (e.g., `dd.json`) leads to a smaller `database.h5` size and so faster loading, although it may lose some additional meta information.
  
- **JSON support  in `load_database` function**
  - The `load_database` can now read `dd.json`, `ini.json`, and `act.json` in addition to existing HDF5 formats (`dd.h5`, `ini.h5`, `act.h5`).

## Example 
- Before this PR: 
  - Test database file: `filtered_only.h5`  (851 MB for 153 cases) (@TimSlendebroek provided)
  - `load_database("filtered_only.h5")` took about 135 seconds.
  
- After this PR:
  - First, convert the original database file into new one with `json` sub-groups: 
     ```julia
     FUSE.sample_and_write_database("filtered_only.h5", "filtered_json.h5"; sampling_ratio=1.01, sub_format=:json)
    ```
   -  new database file: `filtered_json.h5`  (225 MB for 153 cases)
  - `load_database("filtered_json.h5")` took about 30 seconds.

- File size: 851 MB ==> 225 MB 
- Loading time: 135 secs ==> 30 secs

## Future work
- I will implement an additional option in StudyParameters to allow using `sub_format=:json` for creating smaller `database.h5`