@orso82 2025-03-21
@mgyoo86 is this what you were saying that json is actually smaller (and faster it seems) than of HDF5? at least the hierarchical version of HDF5, I guess