@mgyoo86 2025-03-21
@orso82 
Yes, `json` subgroups (`dd.json`) are smaller than `h5` subgroups (`dd.h5`) because the `HDF5` format needs more space to handle its hierarchical structure for many small groups.
For loading speed, there are multiple factors, but I believe the smaller size of the `json` sub-group is beneficial to reducing the loading time.